+++ CHANGELOG.md
@@ -5,6 +5,15 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 [Keep a CHANGELOG](http://keepachangelog.com).
 
 
+## Unreleased
+
+### Added
+
+- RDF-star support for the GraphDB adapter via `gnoa:graphDbRdfStarSupport` manifest property
+
+[Compare v0.1.0...HEAD](https://github.com/rdf-elixir/gno/compare/v0.1.0...HEAD)
+
+
 ## v0.1.0 - 2026-03-18
 
 Initial release